+++ docs/modules/documentation/containers/badge-label/badge-label-header/badge-label-header.component.ts
@@ -1,0 +1,15 @@
+import { Component, OnInit } from '@angular/core';
+
+@Component({
+  selector: 'app-badge-label-header',
+  templateUrl: './badge-label-header.component.html',
+  styleUrls: ['./badge-label-header.component.scss']
+})
+export class BadgeLabelHeaderComponent implements OnInit {
+
+  constructor() { }
+
+  ngOnInit() {
+  }
+
+}