+++ docs/modules/documentation/containers/file-input/file-input-header/file-input-header.component.ts
@@ -1,0 +1,15 @@
+import { Component, OnInit } from '@angular/core';
+
+@Component({
+  selector: 'app-file-input-header',
+  templateUrl: './file-input-header.component.html',
+  styleUrls: ['./file-input-header.component.scss']
+})
+export class FileInputHeaderComponent implements OnInit {
+
+  constructor() { }
+
+  ngOnInit() {
+  }
+
+}