+++ docs/modules/documentation/containers/image/image-header/image-header.component.ts
@@ -1,0 +1,15 @@
+import { Component, OnInit } from '@angular/core';
+
+@Component({
+  selector: 'app-image-header',
+  templateUrl: './image-header.component.html',
+  styleUrls: ['./image-header.component.scss']
+})
+export class ImageHeaderComponent implements OnInit {
+
+  constructor() { }
+
+  ngOnInit() {
+  }
+
+}