+++ docs/modules/documentation/containers/input-group/input-group-header/input-group-header.component.ts
@@ -1,0 +1,15 @@
+import { Component, OnInit } from '@angular/core';
+
+@Component({
+  selector: 'app-input-group-header',
+  templateUrl: './input-group-header.component.html',
+  styleUrls: ['./input-group-header.component.scss']
+})
+export class InputGroupHeaderComponent implements OnInit {
+
+  constructor() { }
+
+  ngOnInit() {
+  }
+
+}