+++ docs/modules/documentation/containers/multi-input/multi-input-header/multi-input-header.component.ts
@@ -1,0 +1,15 @@
+import { Component, OnInit } from '@angular/core';
+
+@Component({
+  selector: 'app-multi-input-header',
+  templateUrl: './multi-input-header.component.html',
+  styleUrls: ['./multi-input-header.component.scss']
+})
+export class MultiInputHeaderComponent implements OnInit {
+
+  constructor() { }
+
+  ngOnInit() {
+  }
+
+}