+++ docs/modules/documentation/containers/panel/panel-docs-header/panel-docs-header.component.ts
@@ -1,0 +1,15 @@
+import { Component, OnInit } from '@angular/core';
+
+@Component({
+  selector: 'app-panel-docs-header',
+  templateUrl: './panel-docs-header.component.html',
+  styleUrls: ['./panel-docs-header.component.scss']
+})
+export class PanelDocsHeaderComponent implements OnInit {
+
+  constructor() { }
+
+  ngOnInit() {
+  }
+
+}