+++ docs/modules/documentation/containers/popover-directive/popover-directive-header/popover-directive-header.component.ts
@@ -1,0 +1,15 @@
+import { Component, OnInit } from '@angular/core';
+
+@Component({
+  selector: 'app-popover-directive-header',
+  templateUrl: './popover-directive-header.component.html',
+  styleUrls: ['./popover-directive-header.component.scss']
+})
+export class PopoverDirectiveHeaderComponent implements OnInit {
+
+  constructor() { }
+
+  ngOnInit() {
+  }
+
+}