+++ docs/modules/documentation/containers/popover/popover-header/popover-header.component.ts
@@ -1,0 +1,15 @@
+import { Component, OnInit } from '@angular/core';
+
+@Component({
+  selector: 'app-popover-header',
+  templateUrl: './popover-header.component.html',
+  styleUrls: ['./popover-header.component.scss']
+})
+export class PopoverHeaderComponent implements OnInit {
+
+  constructor() { }
+
+  ngOnInit() {
+  }
+
+}