+++ docs/modules/documentation/containers/search-input/search-input-header/search-input-header.component.ts
@@ -1,0 +1,15 @@
+import { Component, OnInit } from '@angular/core';
+
+@Component({
+  selector: 'app-search-input-header',
+  templateUrl: './search-input-header.component.html',
+  styleUrls: ['./search-input-header.component.scss']
+})
+export class SearchInputHeaderComponent implements OnInit {
+
+  constructor() { }
+
+  ngOnInit() {
+  }
+
+}