+++ docs/modules/documentation/containers/table/table-docs-header/table-docs-header.component.ts
@@ -1,0 +1,15 @@
+import { Component, OnInit } from '@angular/core';
+
+@Component({
+  selector: 'app-table-docs-header',
+  templateUrl: './table-docs-header.component.html',
+  styleUrls: ['./table-docs-header.component.scss']
+})
+export class TableDocsHeaderComponent implements OnInit {
+
+  constructor() { }
+
+  ngOnInit() {
+  }
+
+}