+++ docs/modules/documentation/containers/time/time-header/time-header.component.ts
@@ -1,0 +1,15 @@
+import { Component, OnInit } from '@angular/core';
+
+@Component({
+  selector: 'app-time-header',
+  templateUrl: './time-header.component.html',
+  styleUrls: ['./time-header.component.scss']
+})
+export class TimeHeaderComponent implements OnInit {
+
+  constructor() { }
+
+  ngOnInit() {
+  }
+
+}