+++ docs/modules/documentation/containers/token/token-header/token-header.component.ts
@@ -1,0 +1,15 @@
+import { Component, OnInit } from '@angular/core';
+
+@Component({
+  selector: 'app-token-header',
+  templateUrl: './token-header.component.html',
+  styleUrls: ['./token-header.component.scss']
+})
+export class TokenHeaderComponent implements OnInit {
+
+  constructor() { }
+
+  ngOnInit() {
+  }
+
+}