+++ CHANGELOG.md
@@ -2,6 +2,31 @@
 
 All notable changes to this project will be documented in this file. See [standard-version](https://github.com/conventional-changelog/standard-version) for commit guidelines.
 
+<a name="0.4.0-rc.8"></a>
+# [0.4.0-rc.8](https://github.com/SAP/fundamental-ngx/compare/v0.4.0-rc.7...v0.4.0-rc.8) (2019-04-08)
+
+
+
+<a name="0.4.0-rc.7"></a>
+# [0.4.0-rc.7](https://github.com/SAP/fundamental-ngx/compare/v0.4.0-rc.6...v0.4.0-rc.7) (2019-04-08)
+
+
+### Features
+
+* allow null values for datetime picker ([#678](https://github.com/SAP/fundamental-ngx/issues/678)) ([ef3dd8a](https://github.com/SAP/fundamental-ngx/commit/ef3dd8a))
+
+
+
+<a name="0.4.0-rc.6"></a>
+# [0.4.0-rc.6](https://github.com/SAP/fundamental-ngx/compare/v0.4.0-rc.5...v0.4.0-rc.6) (2019-04-08)
+
+
+### Bug Fixes
+
+* fixed maxHeight multi-input property ([#686](https://github.com/SAP/fundamental-ngx/issues/686)) ([34facc3](https://github.com/SAP/fundamental-ngx/commit/34facc3))
+
+
+
 <a name="0.4.0-rc.5"></a>
 # [0.4.0-rc.5](https://github.com/SAP/fundamental-ngx/compare/v0.4.0-rc.4...v0.4.0-rc.5) (2019-04-07)